Refactored PyPortal to use PortalBase
